File duration variable changed from int32 to double Added python script to create animations

+++ code/mainwindow.h
@@ -71,7 +71,7 @@ private:
 	
 	double sampleRate;
 	double duration;
-	int fileLength;
+	double fileLength;
 	double xCompensate, yCompensate;
 	double zoom;
